+++ python/instrumentation/openinference-instrumentation-anthropic/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.0.5](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-anthropic-v1.0.4...python-openinference-instrumentation-anthropic-v1.0.5) (2026-05-18)
+
+
+### Bug Fixes
+
+* bump openinference-instrumentation minimum to &gt;=0.1.51 ([#3110](https://github.com/Arize-ai/openinference/issues/3110)) ([bae43ff](https://github.com/Arize-ai/openinference/commit/bae43ff5676fbc4d3a666a15fb3bc50fb73316da))
+
 ## [1.0.4](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-anthropic-v1.0.3...python-openinference-instrumentation-anthropic-v1.0.4) (2026-05-14)