+++ python/instrumentation/openinference-instrumentation-crewai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.7](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-crewai-v1.1.6...python-openinference-instrumentation-crewai-v1.1.7) (2026-05-18)
+
+
+### Bug Fixes
+
+* bump openinference-instrumentation minimum to &gt;=0.1.51 ([#3110](https://github.com/Arize-ai/openinference/issues/3110)) ([bae43ff](https://github.com/Arize-ai/openinference/commit/bae43ff5676fbc4d3a666a15fb3bc50fb73316da))
+
 ## [1.1.6](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-crewai-v1.1.5...python-openinference-instrumentation-crewai-v1.1.6) (2026-05-14)